Chat working over websockets.
The backbone.iobind method seems to be creating a model and saving it
without adding the model to the collection. Only models added by the
server (via socket.broadcast.emit or socket.emit) are added to the
collection. This is useful to assign the id on the server, but means
the model's view isn't rendered until it comes back from the server.