Remove hacky message adding.
After reading the backbone.iobind code it's unclear why binding
socket.io events to a model without an id causes a memory leak,
especially since that model is updated with an id in the RPC
callback. It wouldn't be that difficult to do websocket events
without backbone.iobind, though the Backbone.sync override is nice.